Restore should use pgrep instead of mysqladmin to check for mysql down
For a short window, while mysql is shutting down, it stops accepting
network connections but the mysqld process is still alive.

During this window mysqladmin ping reports that mysql is down and
trying to start mysql fails with:
"mysqld_safe A mysqld process already exists"

Use pgrep (during restore) instead of mysqladmin ping to check mysql
is down to fix this.

Fixes bug: 1197161

Change-Id: Id8b8e15b4b6136c8acbc3cac17691a65a1a4ed14